refactor: one type per file
Mechanical split of every multi-type source file - declarations moved
verbatim, namespaces unchanged (new DispatchRoots/ and StagedPlans/ folders
under Core.Abstractions are physical organization only), per-file usings
pruned, the CS8714 pragma pairs reproduced around exactly the types they
originally spanned.

- GeneratedDispatchRoots.cs -> DispatchRoots/ (13 files)
- StagedPlans.cs -> StagedPlans/ (7 files)
- PipelineExecutors.cs -> 8 executor/gate/cache files in place
- StreamInvoker.cs -> invoker interface + invoker + cache
- BroadcastInvoker.cs -> invoker interface + invoker + cache
- IPipelineExecutor.cs, ExecutionContextScope.cs, StagedPlanModel.cs,
  DescriptorModel.cs -> second types into their own files
