+++ src/Stella.Ergosfare.Core.Abstractions/Context/ExecutionContextScope.cs
@@ -33,13 +33,3 @@ public readonly struct ExecutionContextScope : IDisposable
     /// </summary>
     public void Dispose() => (Context as IPoolReturnable)?.ReturnToPool();
 }
-
-/// <summary>
-/// Implemented by pooled execution contexts; <see cref="ExecutionContextScope.Dispose"/>
-/// and the dispatch paths return contexts through it.
-/// </summary>
-internal interface IPoolReturnable
-{
-    /// <summary>Resets the instance and returns it to its pool.</summary>
-    void ReturnToPool();
-}
