+++ src/Stella.Ergosfare.Core.Abstractions/DispatchRoots/GeneratedDispatchRoots.cs
@@ -163,156 +163,3 @@ public static class GeneratedDispatchRoots
     public static StagedResultPlan? FindStagedResultPlan(Type messageType, Type resultType)
         => StagedResultPlans.TryGetValue((messageType, resultType), out var plan) ? plan : null;
 }
-
-/// <summary>
-/// A dispatch root closed over a concrete message type. A consumer implements
-/// <see cref="IMessageRootVisitor{TReturn, TState}"/> to re-enter a generic context with
-/// the root's type argument and construct its closed dispatch component there — no
-/// reflection involved.
-/// </summary>
-public abstract class MessageRoot
-{
-    /// <summary>Invokes the visitor with this root's message type as the generic argument.</summary>
-    public abstract TReturn Accept<TReturn, TState>(IMessageRootVisitor<TReturn, TState> visitor, TState state);
-}
-
-/// <summary>The concrete closure of <see cref="MessageRoot"/>; instantiated by generated code.</summary>
-public sealed class MessageRoot<TMessage> : MessageRoot where TMessage : IMessage
-{
-    /// <inheritdoc />
-    public override TReturn Accept<TReturn, TState>(IMessageRootVisitor<TReturn, TState> visitor, TState state)
-        => visitor.Visit<TMessage>(state);
-}
-
-/// <summary>Generic re-entry point for consumers of <see cref="MessageRoot"/>.</summary>
-public interface IMessageRootVisitor<out TReturn, in TState>
-{
-    /// <summary>Called with the root's message type as the generic argument.</summary>
-    TReturn Visit<TMessage>(TState state) where TMessage : IMessage;
-}
-
-/// <summary>
-/// A dispatch root closed over a concrete (message, result) pair; see
-/// <see cref="MessageRoot"/>.
-/// </summary>
-public abstract class MessageResultRoot
-{
-    /// <summary>Invokes the visitor with this root's message and result types as the generic arguments.</summary>
-    public abstract TReturn Accept<TReturn, TState>(IMessageResultRootVisitor<TReturn, TState> visitor, TState state);
-}
-
-/// <summary>The concrete closure of <see cref="MessageResultRoot"/>; instantiated by generated code.</summary>
-public sealed class MessageResultRoot<TMessage, TResult> : MessageResultRoot where TMessage : IMessage
-{
-    /// <inheritdoc />
-    public override TReturn Accept<TReturn, TState>(IMessageResultRootVisitor<TReturn, TState> visitor, TState state)
-        => visitor.Visit<TMessage, TResult>(state);
-}
-
-/// <summary>Generic re-entry point for consumers of <see cref="MessageResultRoot"/>.</summary>
-public interface IMessageResultRootVisitor<out TReturn, in TState>
-{
-    /// <summary>Called with the root's message and result types as the generic arguments.</summary>
-    TReturn Visit<TMessage, TResult>(TState state) where TMessage : IMessage;
-}
-
-/// <summary>
-/// A compile-time pipeline plan closed over a void message and its sole async handler;
-/// see <see cref="GeneratedDispatchRoots.AddVoidPlan{TMessage, THandler}"/> and
-/// <see cref="MessageRoot"/> for the visitor re-entry pattern.
-/// </summary>
-public abstract class VoidPlanRoot
-{
-    /// <summary>Invokes the visitor with this plan's message and handler types as the generic arguments.</summary>
-    public abstract TReturn Accept<TReturn, TState>(IVoidPlanRootVisitor<TReturn, TState> visitor, TState state);
-
-    /// <summary>
-    /// The compile-time handler construction path — a <c>Func&lt;THandler&gt;</c> or a
-    /// provider-taking <c>Func&lt;IServiceProvider, THandler&gt;</c>, carried erased and
-    /// cast back inside the executor's closed generic context — or <c>null</c> when the
-    /// generator emitted no factory for the handler.
-    /// </summary>
-    internal virtual object? DirectHandlerFactory => null;
-}
-
-/// <summary>The concrete closure of <see cref="VoidPlanRoot"/>; instantiated by generated code.</summary>
-public sealed class VoidPlanRoot<TMessage, THandler> : VoidPlanRoot
-    where TMessage : notnull, IMessage
-    where THandler : class, IAsyncHandler<TMessage>
-{
-    private readonly object? _directHandlerFactory;
-
-    /// <summary>Creates a plan without a compile-time construction path.</summary>
-    public VoidPlanRoot()
-    {
-    }
-
-    internal VoidPlanRoot(Func<THandler> directHandlerFactory)
-        => _directHandlerFactory = directHandlerFactory;
-
-    internal VoidPlanRoot(Func<IServiceProvider, THandler> directHandlerFactory)
-        => _directHandlerFactory = directHandlerFactory;
-
-    internal override object? DirectHandlerFactory => _directHandlerFactory;
-
-    /// <inheritdoc />
-    public override TReturn Accept<TReturn, TState>(IVoidPlanRootVisitor<TReturn, TState> visitor, TState state)
-        => visitor.Visit<TMessage, THandler>(state);
-}
-
-/// <summary>Generic re-entry point for consumers of <see cref="VoidPlanRoot"/>.</summary>
-public interface IVoidPlanRootVisitor<out TReturn, in TState>
-{
-    /// <summary>Called with the plan's message and handler types as the generic arguments.</summary>
-    TReturn Visit<TMessage, THandler>(TState state)
-        where TMessage : notnull, IMessage
-        where THandler : class, IAsyncHandler<TMessage>;
-}
-
-/// <summary>
-/// A compile-time pipeline plan closed over a result-producing message, its result type
-/// and its sole async handler; the result-producing counterpart of
-/// <see cref="VoidPlanRoot"/>.
-/// </summary>
-public abstract class ResultPlanRoot
-{
-    /// <summary>Invokes the visitor with this plan's message, result and handler types as the generic arguments.</summary>
-    public abstract TReturn Accept<TReturn, TState>(IResultPlanRootVisitor<TReturn, TState> visitor, TState state);
-
-    /// <inheritdoc cref="VoidPlanRoot.DirectHandlerFactory"/>
-    internal virtual object? DirectHandlerFactory => null;
-}
-
-/// <summary>The concrete closure of <see cref="ResultPlanRoot"/>; instantiated by generated code.</summary>
-public sealed class ResultPlanRoot<TMessage, TResult, THandler> : ResultPlanRoot
-    where TMessage : notnull, IMessage
-    where THandler : class, IAsyncHandler<TMessage, TResult>
-{
-    private readonly object? _directHandlerFactory;
-
-    /// <summary>Creates a plan without a compile-time construction path.</summary>
-    public ResultPlanRoot()
-    {
-    }
-
-    internal ResultPlanRoot(Func<THandler> directHandlerFactory)
-        => _directHandlerFactory = directHandlerFactory;
-
-    internal ResultPlanRoot(Func<IServiceProvider, THandler> directHandlerFactory)
-        => _directHandlerFactory = directHandlerFactory;
-
-    internal override object? DirectHandlerFactory => _directHandlerFactory;
-
-    /// <inheritdoc />
-    public override TReturn Accept<TReturn, TState>(IResultPlanRootVisitor<TReturn, TState> visitor, TState state)
-        => visitor.Visit<TMessage, TResult, THandler>(state);
-}
-
-/// <summary>Generic re-entry point for consumers of <see cref="ResultPlanRoot"/>.</summary>
-public interface IResultPlanRootVisitor<out TReturn, in TState>
-{
-    /// <summary>Called with the plan's message, result and handler types as the generic arguments.</summary>
-    TReturn Visit<TMessage, TResult, THandler>(TState state)
-        where TMessage : notnull, IMessage
-        where THandler : class, IAsyncHandler<TMessage, TResult>;
-}
