+++ src/Stella.Ergosfare.Core.Abstractions/IPipelineExecutor.cs
@@ -23,14 +23,3 @@ public interface IPipelineExecutor
     /// <param name="serviceProvider">The provider of the scope the dispatch runs in.</param>
     ValueTask Execute(object message, IExecutionContext context, IServiceProvider serviceProvider);
 }
-
-/// <summary>
-/// A result-producing message pipeline closed over its concrete message type; see
-/// <see cref="IPipelineExecutor"/>.
-/// </summary>
-/// <typeparam name="TResult">The result type produced by the pipeline.</typeparam>
-public interface IPipelineExecutor<TResult>
-{
-    /// <inheritdoc cref="IPipelineExecutor.Execute"/>
-    ValueTask<TResult> Execute(object message, IExecutionContext context, IServiceProvider serviceProvider);
-}
