+++ src/Stella.Ergosfare.Core/Internal/Mediator/PipelineExecutorCache.cs
@@ -7,534 +7,8 @@ using Stella.Ergosfare.Core.Abstractions.Factories;
 using Stella.Ergosfare.Core.Abstractions.Registry.Descriptors;
 using Stella.Ergosfare.Core.Abstractions.Handlers;
 using Stella.Ergosfare.Core.Abstractions.Strategies;
-using Stella.Ergosfare.Core.Internal.Factories;
 
 namespace Stella.Ergosfare.Core.Internal.Mediator;
-
-/// <summary>
-/// Result-producing pipeline closed over the concrete <typeparamref name="TMessage"/>.
-/// Because <typeparamref name="TMessage"/> is the message's runtime type here, the mediation
-/// strategy's typed seam always hits — the handler is invoked through its typed member and
-/// its <see cref="ValueTask{TResult}"/> never crosses an object-typed bridge.
-/// </summary>
-#pragma warning disable CS8714 // TResult is used as a pattern type argument; handler contracts declare notnull results
-internal sealed class ResultPipelineExecutor<TMessage, TResult>(
-    IMessageDescriptor descriptor,
-    IMessageDependenciesFactory dependenciesFactory,
-    IResultAdapterService? resultAdapterService,
-    string[] groups) : IPipelineExecutor<TResult>
-    where TMessage : notnull
-{
-    private readonly SingleAsyncHandlerMediationStrategy<TMessage, TResult> _strategy = new(resultAdapterService);
-
-    // Adapter service split by shape: the concrete service can report emptiness cheaply,
-    // a foreign implementation always routes through the strategy (which consults it).
-    private readonly ResultAdapterService? _concreteAdapters = resultAdapterService as ResultAdapterService;
-    private readonly bool _foreignAdapters = resultAdapterService is not null and not ResultAdapterService;
-
-    // Dependencies cached per executor (executors are already per message type + groups),
-    // re-validated against the registry version — turns the per-dispatch factory call and
-    // cache lookup into a single field read + version compare.
-    private IMessageDependencies? _cachedDependencies;
-    private MessageDependencies? _cachedFastDependencies;
-    private int _cachedVersion = int.MinValue;
-
-    public ValueTask<TResult> Execute(object message, IExecutionContext context, IServiceProvider serviceProvider)
-    {
-        var dependencies = GetDependencies();
-
-        // Zero-interceptor, single-handler, no-adapter dispatch: invoke the handler's typed
-        // member directly and hand its ValueTask straight back — no async state machine,
-        // no interface-dispatched Count checks. Mirrors the strategy's fast path exactly.
-        if (_cachedFastDependencies?.FastSingleHandler is { } handlerReference
-            && !_foreignAdapters
-            && (_concreteAdapters is null || _concreteAdapters.IsEmpty))
-        {
-            var handler = handlerReference.Resolve(serviceProvider);
-
-            switch (handler)
-            {
-                case IAsyncHandler<TMessage, TResult> asyncHandler:
-                    return asyncHandler.HandleAsync((TMessage)message, context);
-                case IHandler<TMessage, ValueTask<TResult>> valueTaskShaped:
-                    return valueTaskShaped.Handle((TMessage)message, context);
-                case IHandler<TMessage, TResult> syncHandler:
-                    return ValueTask.FromResult(syncHandler.Handle((TMessage)message, context));
-            }
-
-            // Unsupported handler contract: fall through so the strategy raises its
-            // canonical NotSupportedException.
-        }
-
-        return _strategy.Mediate((TMessage)message, dependencies, context, serviceProvider);
-    }
-
-    private IMessageDependencies GetDependencies()
-    {
-        if (dependenciesFactory is MessageDependenciesFactory typedFactory)
-        {
-            var cached = _cachedDependencies;
-
-            if (cached is not null && _cachedVersion == typedFactory.CurrentRegistryVersion)
-            {
-                return cached;
-            }
-
-            // Create runs the registry-version invalidation and rebuilds; races are benign —
-            // both writers publish equivalent, idempotent state.
-            var dependencies = typedFactory.Create(typeof(TMessage), descriptor, groups);
-            _cachedFastDependencies = dependencies as MessageDependencies;
-            _cachedDependencies = dependencies;
-            _cachedVersion = typedFactory.CurrentRegistryVersion;
-            return dependencies;
-        }
-
-        // Foreign factory implementations keep the original per-dispatch behavior.
-        return dependenciesFactory.Create(typeof(TMessage), descriptor, groups);
-    }
-}
-
-#pragma warning restore CS8714
-
-/// <summary>
-/// Void pipeline closed over the concrete <typeparamref name="TMessage"/>; see
-/// <see cref="ResultPipelineExecutor{TMessage, TResult}"/>.
-/// </summary>
-internal sealed class VoidPipelineExecutor<TMessage>(
-    IMessageDescriptor descriptor,
-    IMessageDependenciesFactory dependenciesFactory,
-    IResultAdapterService? resultAdapterService,
-    string[] groups) : IPipelineExecutor
-    where TMessage : IMessage
-{
-    private readonly SingleAsyncHandlerMediationStrategy<TMessage> _strategy = new(resultAdapterService);
-
-    private readonly ResultAdapterService? _concreteAdapters = resultAdapterService as ResultAdapterService;
-    private readonly bool _foreignAdapters = resultAdapterService is not null and not ResultAdapterService;
-
-    private IMessageDependencies? _cachedDependencies;
-    private MessageDependencies? _cachedFastDependencies;
-    private int _cachedVersion = int.MinValue;
-
-    public ValueTask Execute(object message, IExecutionContext context, IServiceProvider serviceProvider)
-    {
-        var dependencies = GetDependencies();
-
-        if (_cachedFastDependencies?.FastSingleHandler is { } handlerReference
-            && !_foreignAdapters
-            && (_concreteAdapters is null || _concreteAdapters.IsEmpty))
-        {
-            var handler = handlerReference.Resolve(serviceProvider);
-
-            switch (handler)
-            {
-                case IAsyncHandler<TMessage> asyncHandler:
-                    return asyncHandler.HandleAsync((TMessage)message, context);
-                case IHandler<TMessage, ValueTask> valueTaskShaped:
-                    return valueTaskShaped.Handle((TMessage)message, context);
-                case IHandler<TMessage, object> syncHandler:
-                    syncHandler.Handle((TMessage)message, context);
-                    return ValueTask.CompletedTask;
-            }
-        }
-
-        return _strategy.Mediate((TMessage)message, dependencies, context, serviceProvider);
-    }
-
-    private IMessageDependencies GetDependencies()
-    {
-        if (dependenciesFactory is MessageDependenciesFactory typedFactory)
-        {
-            var cached = _cachedDependencies;
-
-            if (cached is not null && _cachedVersion == typedFactory.CurrentRegistryVersion)
-            {
-                return cached;
-            }
-
-            var dependencies = typedFactory.Create(typeof(TMessage), descriptor, groups);
-            _cachedFastDependencies = dependencies as MessageDependencies;
-            _cachedDependencies = dependencies;
-            _cachedVersion = typedFactory.CurrentRegistryVersion;
-            return dependencies;
-        }
-
-        return dependenciesFactory.Create(typeof(TMessage), descriptor, groups);
-    }
-}
-
-/// <summary>
-/// Void pipeline closed over both the message and its compile-time-known sole handler —
-/// the executor a generated void plan constructs. The fast path resolves the handler
-/// reference exactly like <see cref="VoidPipelineExecutor{TMessage}"/> but invokes it
-/// through the closed <typeparamref name="THandler"/> type, so the call devirtualizes
-/// (and inlines for sealed handlers) instead of walking the contract pattern match. The
-/// plan is advisory: the same registry-version-guarded dependency cache re-validates the
-/// pipeline, and any mismatch — interceptors registered at runtime, a differently-typed
-/// handler instance, configured adapters — falls back to the runtime dispatch shape,
-/// preserving semantics exactly.
-/// </summary>
-internal sealed class GeneratedVoidPipelineExecutor<TMessage, THandler>(
-    IMessageDescriptor descriptor,
-    IMessageDependenciesFactory dependenciesFactory,
-    IResultAdapterService? resultAdapterService,
-    string[] groups,
-    Func<THandler>? directHandlerFactory = null,
-    Func<IServiceProvider, THandler>? providerHandlerFactory = null) : IPipelineExecutor
-    where TMessage : notnull, IMessage
-    where THandler : class, IAsyncHandler<TMessage>
-{
-    private readonly SingleAsyncHandlerMediationStrategy<TMessage> _strategy = new(resultAdapterService);
-
-    private readonly ResultAdapterService? _concreteAdapters = resultAdapterService as ResultAdapterService;
-    private readonly bool _foreignAdapters = resultAdapterService is not null and not ResultAdapterService;
-
-    // Compile-time construction paths for the planned handler; discarded up front for
-    // disposable handlers — the container tracks transient disposables in the resolving
-    // scope, direct construction would not. The provider-taking shape covers handlers
-    // with constructor dependencies: it resolves them from the dispatching scope's
-    // provider, exactly where container activation would resolve them.
-    private static readonly bool HandlerIsDisposable =
-        typeof(IDisposable).IsAssignableFrom(typeof(THandler)) || typeof(IAsyncDisposable).IsAssignableFrom(typeof(THandler));
-
-    private readonly Func<THandler>? _directHandlerFactory = HandlerIsDisposable ? null : directHandlerFactory;
-    private readonly Func<IServiceProvider, THandler>? _providerHandlerFactory = HandlerIsDisposable ? null : providerHandlerFactory;
-
-    private IMessageDependencies? _cachedDependencies;
-    private MessageDependencies? _cachedFastDependencies;
-    private int _cachedVersion = int.MinValue;
-
-    // Re-validated with the dependency cache: true only while the registry's sole handler
-    // is the planned type, instances are not memoized, and the handler's effective DI
-    // registration is the module's own plain transient one — the exact conditions under
-    // which GetRequiredService is observably nothing but a constructor call.
-    private bool _useDirectConstruction;
-
-    public ValueTask Execute(object message, IExecutionContext context, IServiceProvider serviceProvider)
-    {
-        var dependencies = GetDependencies();
-
-        if (_cachedFastDependencies?.FastSingleHandler is { } handlerReference
-            && !_foreignAdapters
-            && (_concreteAdapters is null || _concreteAdapters.IsEmpty))
-        {
-            // The handler-type re-check pins the racy flag to the reference actually in
-            // hand: a version transition observed halfway can only route back through the
-            // container, never construct a type the registry no longer plans.
-            IHandler handler = _useDirectConstruction && handlerReference.HandlerType == typeof(THandler)
-                ? _directHandlerFactory is not null ? _directHandlerFactory() : _providerHandlerFactory!(serviceProvider)
-                : handlerReference.Resolve(serviceProvider);
-
-            // The compile-time plan's handler type: a devirtualized call, no pattern
-            // match. A runtime re-registration can put a differently-typed handler here;
-            // the contract switch below then dispatches it exactly as the runtime
-            // executor would.
-            if (handler is THandler planned)
-            {
-                return planned.HandleAsync((TMessage)message, context);
-            }
-
-            switch (handler)
-            {
-                case IAsyncHandler<TMessage> asyncHandler:
-                    return asyncHandler.HandleAsync((TMessage)message, context);
-                case IHandler<TMessage, ValueTask> valueTaskShaped:
-                    return valueTaskShaped.Handle((TMessage)message, context);
-                case IHandler<TMessage, object> syncHandler:
-                    syncHandler.Handle((TMessage)message, context);
-                    return ValueTask.CompletedTask;
-            }
-        }
-
-        return _strategy.Mediate((TMessage)message, dependencies, context, serviceProvider);
-    }
-
-    private IMessageDependencies GetDependencies()
-    {
-        if (dependenciesFactory is MessageDependenciesFactory typedFactory)
-        {
-            var cached = _cachedDependencies;
-
-            if (cached is not null && _cachedVersion == typedFactory.CurrentRegistryVersion)
-            {
-                return cached;
-            }
-
-            var dependencies = typedFactory.Create(typeof(TMessage), descriptor, groups);
-            var fastDependencies = dependencies as MessageDependencies;
-            _cachedFastDependencies = fastDependencies;
-            _cachedDependencies = dependencies;
-            _useDirectConstruction = (_directHandlerFactory is not null || _providerHandlerFactory is not null)
-                && fastDependencies is { MemoizedInstances: false, FastSingleHandler.HandlerType: var plannedType }
-                && plannedType == typeof(THandler)
-                && typedFactory.IsPlainTransientRegistration(typeof(THandler));
-            _cachedVersion = typedFactory.CurrentRegistryVersion;
-            return dependencies;
-        }
-
-        _useDirectConstruction = false;
-        return dependenciesFactory.Create(typeof(TMessage), descriptor, groups);
-    }
-}
-
-/// <summary>
-/// Result-producing counterpart of
-/// <see cref="GeneratedVoidPipelineExecutor{TMessage, THandler}"/>: closed over the
-/// message, its result and its compile-time-known sole async handler, so the handler call
-/// devirtualizes instead of walking the contract pattern match. The same advisory-plan
-/// contract applies — the registry-version-guarded dependency cache re-validates the
-/// pipeline and any mismatch falls back to the runtime dispatch shape.
-/// </summary>
-#pragma warning disable CS8714 // TResult is used as a pattern type argument; handler contracts declare notnull results
-internal sealed class GeneratedResultPipelineExecutor<TMessage, TResult, THandler>(
-    IMessageDescriptor descriptor,
-    IMessageDependenciesFactory dependenciesFactory,
-    IResultAdapterService? resultAdapterService,
-    string[] groups,
-    Func<THandler>? directHandlerFactory = null,
-    Func<IServiceProvider, THandler>? providerHandlerFactory = null) : IPipelineExecutor<TResult>
-    where TMessage : notnull, IMessage
-    where THandler : class, IAsyncHandler<TMessage, TResult>
-{
-    private readonly SingleAsyncHandlerMediationStrategy<TMessage, TResult> _strategy = new(resultAdapterService);
-
-    private readonly ResultAdapterService? _concreteAdapters = resultAdapterService as ResultAdapterService;
-    private readonly bool _foreignAdapters = resultAdapterService is not null and not ResultAdapterService;
-
-    private static readonly bool HandlerIsDisposable =
-        typeof(IDisposable).IsAssignableFrom(typeof(THandler)) || typeof(IAsyncDisposable).IsAssignableFrom(typeof(THandler));
-
-    private readonly Func<THandler>? _directHandlerFactory = HandlerIsDisposable ? null : directHandlerFactory;
-    private readonly Func<IServiceProvider, THandler>? _providerHandlerFactory = HandlerIsDisposable ? null : providerHandlerFactory;
-
-    private IMessageDependencies? _cachedDependencies;
-    private MessageDependencies? _cachedFastDependencies;
-    private int _cachedVersion = int.MinValue;
-    private bool _useDirectConstruction;
-
-    public ValueTask<TResult> Execute(object message, IExecutionContext context, IServiceProvider serviceProvider)
-    {
-        var dependencies = GetDependencies();
-
-        if (_cachedFastDependencies?.FastSingleHandler is { } handlerReference
-            && !_foreignAdapters
-            && (_concreteAdapters is null || _concreteAdapters.IsEmpty))
-        {
-            IHandler handler = _useDirectConstruction && handlerReference.HandlerType == typeof(THandler)
-                ? _directHandlerFactory is not null ? _directHandlerFactory() : _providerHandlerFactory!(serviceProvider)
-                : handlerReference.Resolve(serviceProvider);
-
-            if (handler is THandler planned)
-            {
-                return planned.HandleAsync((TMessage)message, context);
-            }
-
-            switch (handler)
-            {
-                case IAsyncHandler<TMessage, TResult> asyncHandler:
-                    return asyncHandler.HandleAsync((TMessage)message, context);
-                case IHandler<TMessage, ValueTask<TResult>> valueTaskShaped:
-                    return valueTaskShaped.Handle((TMessage)message, context);
-                case IHandler<TMessage, TResult> syncHandler:
-                    return ValueTask.FromResult(syncHandler.Handle((TMessage)message, context));
-            }
-        }
-
-        return _strategy.Mediate((TMessage)message, dependencies, context, serviceProvider);
-    }
-
-    private IMessageDependencies GetDependencies()
-    {
-        if (dependenciesFactory is MessageDependenciesFactory typedFactory)
-        {
-            var cached = _cachedDependencies;
-
-            if (cached is not null && _cachedVersion == typedFactory.CurrentRegistryVersion)
-            {
-                return cached;
-            }
-
-            var dependencies = typedFactory.Create(typeof(TMessage), descriptor, groups);
-            var fastDependencies = dependencies as MessageDependencies;
-            _cachedFastDependencies = fastDependencies;
-            _cachedDependencies = dependencies;
-            _useDirectConstruction = (_directHandlerFactory is not null || _providerHandlerFactory is not null)
-                && fastDependencies is { MemoizedInstances: false, FastSingleHandler.HandlerType: var plannedType }
-                && plannedType == typeof(THandler)
-                && typedFactory.IsPlainTransientRegistration(typeof(THandler));
-            _cachedVersion = typedFactory.CurrentRegistryVersion;
-            return dependencies;
-        }
-
-        _useDirectConstruction = false;
-        return dependenciesFactory.Create(typeof(TMessage), descriptor, groups);
-    }
-}
-#pragma warning restore CS8714
-
-/// <summary>
-/// The staged plans' advisory gate: whether the live pipeline is exactly the composition a
-/// plan was baked against — one main handler of the planned type and the four interceptor
-/// stages matching the planned type lists in order. Anything else (a runtime-registered
-/// interceptor, a different or additional handler, reordered stages) fails the match and
-/// keeps the dispatch on the runtime strategy.
-/// </summary>
-internal static class StagedPlanGate
-{
-    internal static bool Matches(MessageDependencies dependencies, StagedPlanComposition composition)
-        => dependencies.Handlers.Count == 1
-           && dependencies.Handlers[0].HandlerType == composition.HandlerType
-           && StageMatches(dependencies.PreInterceptors, composition.PreInterceptorTypeArray)
-           && StageMatches(dependencies.PostInterceptors, composition.PostInterceptorTypeArray)
-           && StageMatches(dependencies.ExceptionInterceptors, composition.ExceptionInterceptorTypeArray)
-           && StageMatches(dependencies.FinalInterceptors, composition.FinalInterceptorTypeArray);
-
-    private static bool StageMatches<THandler, TDescriptor>(
-        IReadOnlyList<IHandlerReference<THandler, TDescriptor>> stage,
-        Type[] baked)
-        where TDescriptor : IHandlerDescriptor
-    {
-        if (stage.Count != baked.Length)
-        {
-            return false;
-        }
-
-        for (var i = 0; i < baked.Length; i++)
-        {
-            if (stage[i].HandlerType != baked[i])
-            {
-                return false;
-            }
-        }
-
-        return true;
-    }
-}
-
-/// <summary>
-/// Void pipeline hosting a staged plan: bespoke straight-line code for the message's whole
-/// interceptor-bearing pipeline. The plan is advisory — the registry-version-guarded
-/// dependency cache re-validates the plan's composition against the live pipeline
-/// (<see cref="StagedPlanGate"/>) plus the memoization and adapter gates, and any mismatch
-/// falls back to the runtime strategy, preserving semantics exactly. The plan resolves its
-/// participants from the dispatching scope's provider — outside memoized mode that is
-/// literally what the runtime handler references do, so container semantics are preserved.
-/// Like every version-guarded executor, a dispatch racing a registration may run the
-/// previous shape once; it never runs a shape the registry has not published.
-/// </summary>
-internal sealed class StagedVoidPipelineExecutor<TMessage>(
-    IMessageDescriptor descriptor,
-    IMessageDependenciesFactory dependenciesFactory,
-    IResultAdapterService? resultAdapterService,
-    string[] groups,
-    StagedVoidPlan<TMessage> plan) : IPipelineExecutor
-    where TMessage : notnull, IMessage
-{
-    private readonly SingleAsyncHandlerMediationStrategy<TMessage> _strategy = new(resultAdapterService);
-
-    private readonly ResultAdapterService? _concreteAdapters = resultAdapterService as ResultAdapterService;
-    private readonly bool _foreignAdapters = resultAdapterService is not null and not ResultAdapterService;
-
-    private IMessageDependencies? _cachedDependencies;
-    private int _cachedVersion = int.MinValue;
-    private bool _useStagedPlan;
-
-    public ValueTask Execute(object message, IExecutionContext context, IServiceProvider serviceProvider)
-    {
-        var dependencies = GetDependencies();
-
-        if (_useStagedPlan)
-        {
-            return plan.Execute((TMessage)message, context, serviceProvider);
-        }
-
-        return _strategy.Mediate((TMessage)message, dependencies, context, serviceProvider);
-    }
-
-    private IMessageDependencies GetDependencies()
-    {
-        if (dependenciesFactory is MessageDependenciesFactory typedFactory)
-        {
-            var cached = _cachedDependencies;
-
-            if (cached is not null && _cachedVersion == typedFactory.CurrentRegistryVersion)
-            {
-                return cached;
-            }
-
-            var dependencies = typedFactory.Create(typeof(TMessage), descriptor, groups);
-            _cachedDependencies = dependencies;
-            _useStagedPlan = !_foreignAdapters
-                && (_concreteAdapters is null || _concreteAdapters.IsEmpty)
-                && dependencies is MessageDependencies { MemoizedInstances: false } fastDependencies
-                && StagedPlanGate.Matches(fastDependencies, plan.Composition);
-            _cachedVersion = typedFactory.CurrentRegistryVersion;
-            return dependencies;
-        }
-
-        _useStagedPlan = false;
-        return dependenciesFactory.Create(typeof(TMessage), descriptor, groups);
-    }
-}
-
-/// <summary>
-/// Result-producing counterpart of <see cref="StagedVoidPipelineExecutor{TMessage}"/>;
-/// the same advisory contract and gates apply.
-/// </summary>
-internal sealed class StagedResultPipelineExecutor<TMessage, TResult>(
-    IMessageDescriptor descriptor,
-    IMessageDependenciesFactory dependenciesFactory,
-    IResultAdapterService? resultAdapterService,
-    string[] groups,
-    StagedResultPlan<TMessage, TResult> plan) : IPipelineExecutor<TResult>
-    where TMessage : notnull, IMessage
-{
-    private readonly SingleAsyncHandlerMediationStrategy<TMessage, TResult> _strategy = new(resultAdapterService);
-
-    private readonly ResultAdapterService? _concreteAdapters = resultAdapterService as ResultAdapterService;
-    private readonly bool _foreignAdapters = resultAdapterService is not null and not ResultAdapterService;
-
-    private IMessageDependencies? _cachedDependencies;
-    private int _cachedVersion = int.MinValue;
-    private bool _useStagedPlan;
-
-    public ValueTask<TResult> Execute(object message, IExecutionContext context, IServiceProvider serviceProvider)
-    {
-        var dependencies = GetDependencies();
-
-        if (_useStagedPlan)
-        {
-            return plan.Execute((TMessage)message, context, serviceProvider);
-        }
-
-        return _strategy.Mediate((TMessage)message, dependencies, context, serviceProvider);
-    }
-
-    private IMessageDependencies GetDependencies()
-    {
-        if (dependenciesFactory is MessageDependenciesFactory typedFactory)
-        {
-            var cached = _cachedDependencies;
-
-            if (cached is not null && _cachedVersion == typedFactory.CurrentRegistryVersion)
-            {
-                return cached;
-            }
-
-            var dependencies = typedFactory.Create(typeof(TMessage), descriptor, groups);
-            _cachedDependencies = dependencies;
-            _useStagedPlan = !_foreignAdapters
-                && (_concreteAdapters is null || _concreteAdapters.IsEmpty)
-                && dependencies is MessageDependencies { MemoizedInstances: false } fastDependencies
-                && StagedPlanGate.Matches(fastDependencies, plan.Composition);
-            _cachedVersion = typedFactory.CurrentRegistryVersion;
-            return dependencies;
-        }
-
-        _useStagedPlan = false;
-        return dependenciesFactory.Create(typeof(TMessage), descriptor, groups);
-    }
-}
 
 /// <summary>
 /// Process-wide cache of pipeline executors, one per (message runtime type, result type,
