+++ src/Stella.Ergosfare.Events/EventBroadcastInvoker[TEvent].cs
@@ -1,5 +1,3 @@
-using System.Collections.Concurrent;
-using System.Diagnostics.CodeAnalysis;
 using System.Runtime.CompilerServices;
 using Stella.Ergosfare.Core;
 using Stella.Ergosfare.Core.Abstractions;
@@ -11,37 +9,6 @@ using Stella.Ergosfare.Core.Internal.Mediator;
 using Stella.Ergosfare.Events.Abstractions;
 
 namespace Stella.Ergosfare.Events;
-
-/// <summary>
-/// Publishes an event through a pipeline closed over the event's concrete type, so broadcast
-/// handlers are always invoked through their typed members — interface-erased publishes
-/// (<c>PublishAsync((IEvent)e)</c>) resolve the invoker from the event's runtime type.
-/// Invokers are closed once per event type and cached; the per-call
-/// <see cref="EventMediationSettings"/> flows into a fresh strategy instance, as before.
-/// </summary>
-internal interface IEventBroadcastInvoker
-{
-    ValueTask Publish(object @event, EventMediationSettings? settings, CancellationToken cancellationToken,
-        IMessageMediator mediator, ActualTypeOrFirstAssignableTypeMessageResolveStrategy resolveStrategy,
-        IResultAdapterService? resultAdapterService, IExecutionContext? externalContext = null,
-        IEnumerable<string>? groupsOverride = null);
-
-    /// <summary>
-    /// Engine-backed publish: the concrete dispatch machinery is known by construction, so
-    /// the publish runs directly against the engine's plan and the caller's scope
-    /// provider — grouped filters included, resolved from the grouped plan slot. External
-    /// contexts resolve the scope's <see cref="IMessageMediator"/> on demand and run the
-    /// original overload — semantics unchanged.
-    /// <paramref name="groupsOverride"/> carries a facade-level group filter (a
-    /// <see cref="GroupSet"/>) without a settings object; when present it takes
-    /// precedence over the settings' groups.
-    /// </summary>
-    ValueTask Publish(object @event, EventMediationSettings? settings, CancellationToken cancellationToken,
-        MessageDispatchEngine engine, IServiceProvider serviceProvider,
-        ActualTypeOrFirstAssignableTypeMessageResolveStrategy resolveStrategy,
-        IResultAdapterService? resultAdapterService, IExecutionContext? externalContext = null,
-        IEnumerable<string>? groupsOverride = null);
-}
 
 internal sealed class EventBroadcastInvoker<TEvent> : IEventBroadcastInvoker
     where TEvent : notnull
@@ -462,63 +429,5 @@ internal sealed class EventBroadcastInvoker<TEvent> : IEventBroadcastInvoker
                          ?? throw new NoHandlerFoundException(typeof(TEvent));
 
         return factory.Create(typeof(TEvent), descriptor, groups);
-    }
-}
-
-/// <summary>
-/// Process-wide cache of <see cref="IEventBroadcastInvoker"/> instances, one per event
-/// runtime type — one <see cref="Type.MakeGenericType"/> per event type.
-/// </summary>
-internal static class EventBroadcastInvokerCache
-{
-    private static readonly ConcurrentDictionary<Type, IEventBroadcastInvoker> Invokers = new();
-
-    /// <summary>
-    /// Static-generic view of the cache for callers that know the event's concrete type at
-    /// compile time: the invoker resolves once per closed type into a static readonly
-    /// field, so the typed publish overload skips the per-call dictionary lookup. Shares
-    /// the dictionary's instance, keeping the plan cache one-per-event-type (and
-    /// factory-keyed, so container isolation is unchanged). Callers must guard with
-    /// <c>@event.GetType() == typeof(TEvent)</c> — a base-typed generic call must keep
-    /// resolving by the runtime type.
-    /// </summary>
-    internal static class Holder<TEvent> where TEvent : notnull
-    {
-        public static readonly IEventBroadcastInvoker Instance = Get(typeof(TEvent));
-    }
-
-    [UnconditionalSuppressMessage("Trimming", "IL2055",
-        Justification = "The invoker generic is closed over a live event's runtime type; the event roots its type.")]
-    [UnconditionalSuppressMessage("AOT", "IL3050",
-        Justification = "Generated dispatch roots cover source-generated event types; this reflective path is the " +
-                        "JIT fallback for runtime-only registrations.")]
-    public static IEventBroadcastInvoker Get(Type eventType)
-    {
-        if (Invokers.TryGetValue(eventType, out var invoker))
-        {
-            return invoker;
-        }
-
-        // Generated dispatch roots close the invoker generic at compile time; the
-        // reflective path below only serves event types without a root.
-        if (GeneratedDispatchRoots.FindMessage(eventType) is { } root)
-        {
-            return Invokers.GetOrAdd(eventType, root.Accept(InvokerVisitor.Instance, state: false));
-        }
-
-        return Invokers.GetOrAdd(eventType,
-            static t => (IEventBroadcastInvoker)Activator.CreateInstance(typeof(EventBroadcastInvoker<>).MakeGenericType(t))!);
-    }
-
-    /// <summary>
-    /// Re-enters a generic context with a root's event type and constructs the closed
-    /// broadcast invoker there — no <see cref="Type.MakeGenericType"/>, no reflection.
-    /// </summary>
-    private sealed class InvokerVisitor : IMessageRootVisitor<IEventBroadcastInvoker, bool>
-    {
-        public static readonly InvokerVisitor Instance = new();
-
-        public IEventBroadcastInvoker Visit<TMessage>(bool state) where TMessage : IMessage
-            => new EventBroadcastInvoker<TMessage>();
     }
 }
