+++ src/Stella.Ergosfare.SourceGenerator/Models/StagedPlanModel.cs
@@ -3,26 +3,6 @@ using System.Collections.Immutable;
 namespace Stella.Ergosfare.SourceGenerator.Models;
 
 /// <summary>
-///     The pattern-match arm the runtime invocation strategy would select for one staged
-///     interceptor call — determined at compile time from the interceptor's implemented
-///     contracts, so the emitted call invokes exactly the member the strategy would.
-/// </summary>
-internal enum StagedCallArm : byte
-{
-    /// <summary>The result-typed asynchronous contract (first arm of the runtime pattern match).</summary>
-    AsyncTyped,
-
-    /// <summary>The result-agnostic asynchronous contract (second arm; also the sole async pre arm).</summary>
-    AsyncAgnostic,
-
-    /// <summary>The synchronous contract (last arm).</summary>
-    Sync,
-}
-
-/// <summary>One staged interceptor call: the concrete type to resolve and the arm to invoke it through.</summary>
-internal readonly record struct StagedCallModel(string TypeExpression, StagedCallArm Arm);
-
-/// <summary>
 ///     A staged pipeline plan ready for emission: a message whose whole discovered
 ///     pipeline — one async handler plus at least one interceptor stage — could be
 ///     modeled exactly, with every stage in the runtime shape-builder's execution order
